+++ src/CSSmodule/navbar.css
@@ -1,0 +1,9 @@
+.header{
+    width: 100%;
+    height: 50px;
+}
+
+.nav{
+    display: flex;
+    
+}
