--- src/components/pages/home.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-const Homepage = () => {
-    return(
-        <main>
-            <p>Hello world</p>
-        </main>
-    )
-}
-
-export default Homepage;
