+++ src/pages/home.js
@@ -1,0 +1,8 @@
+import React from 'react';
+
+const Homepage = () => (
+  <main>
+  </main>
+);
+
+export default Homepage;
