+++ src/redux/reducer/currencies.js
@@ -1,0 +1,22 @@
+import { createSlice } from '@reduxjs/toolKit';
+import fetchCurrency from './coinAction';
+
+const initialState = {
+  currencies: [],
+  status: null,
+};
+
+const currencyReducer = createSlice({
+  name: 'currency',
+  initialState,
+  extraReducers: (builder) => {
+    builder.addCase(fetchCurrency.fulfilled, (state, action) => {
+      const items = state;
+      items.status = 'success';
+      items.currencies = action.payload;
+      console.log(items.currencies);
+    });
+  },
+});
+
+export default currencyReducer.reducer;
